--- __main__.py
@@ -1,13 +1,0 @@
-"""
-DIMet scripts are in src/ folder
-this __main__ file redirects towards it
-
-johaGL 2022
-"""
-from src import dimet_message
-
-print(dimet_message())
-# src/__init__ file continues the steps
-
-
-
