+++ ancient_src/differential_univariate.py
