+++ ancient_src/functions_diffmode.py
