+++ ancient_src/isotopologcontrib_stacked.py
