+++ ancient_src/metabologram.py
