+++ ancient_src/old_isotopologcontrib.py
