--- drafts/test_main.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-
-
-import src.main as main
-
-
-def test_square():
-    assert main.square(4) == 16
