+++ src/malito.py
@@ -1,0 +1,2 @@
+def nomejodas():
+    return "nomejodas"
